New version

+++ lambda/combine_files.py
@@ -1,46 +1,45 @@
 import importlib
 import os
 import util
+from database import Database, Entry
+from typing import Any, Dict, List
 
 
-def combine(bucket_name, key, input_format, output_format, offsets, params):
-  util.print_read(input_format, key, params)
-
-  batch_size = params["batch_size"] if "batch_size" in params else 1
-  file_id = int((input_format["file_id"] + batch_size - 1) / batch_size)
-
-  # TODO: Fix.
-  if "batch_size" in params:
-    output_format["file_id"] = file_id
-  else:
-    output_format["file_id"] = input_format["bin"]
+def combine(database: Database, table_name, key, input_format, output_format, offsets, params):
+  output_format["file_id"] = input_format["bin"]
   output_format["bin"] = 1
+  output_format["num_bins"] = 1
+  output_format["num_files"] = input_format["num_bins"]
+  file_name = util.file_name(output_format)
   util.make_folder(output_format)
-
-  [combine, keys, last] = util.combine_instance(bucket_name, key, params)
+  [combine, last_file, keys] = util.combine_instance(table_name, key, params)
   if combine:
-    if "batch_size" in params:
-      output_format["last"] = last
-    else:
-      output_format["last"] = (output_format["file_id"] == params["num_bins"])
     msg = "Combining TIMESTAMP {0:f} NONCE {1:d} BIN {2:d} FILE {3:d}"
     msg = msg.format(input_format["timestamp"], input_format["nonce"], input_format["bin"], input_format["file_id"])
     print(msg)
 
-    format_lib = importlib.import_module(params["format"])
-    iterator = getattr(format_lib, "Iterator")
-    file_name = util.file_name(output_format)
+    format_lib = importlib.import_module(params["output_format"])
+    iterator_class = getattr(format_lib, "Iterator")
     temp_name = "/tmp/{0:s}".format(file_name)
     # Make this deterministic and combine in the same order
     keys.sort()
-    iterator.combine(bucket_name, keys, temp_name, params)
-    f = open(temp_name, "rb")
-    util.write(input_format, params["bucket"], file_name, f, params)
-    f.close()
+    entries: List[Entry] = list(map(lambda key: database.get_entry(table_name, key), keys))
+    metadata: Dict[str, str] = {}
+    if database.contains(table_name, file_name):
+      return True
+
+    with open(temp_name, "wb+") as f:
+      metadata = iterator_class.combine(entries, f, params)
+
+    found = database.contains(table_name, file_name)
+    if not found:
+      with open(temp_name, "rb") as f:
+        database.put(params["bucket"], file_name, f, metadata)
     os.remove(temp_name)
+    return True
+  else:
+    return database.contains(table_name, file_name) or key != last_file
 
 
 def handler(event, context):
-  [bucket_name, key, params] = util.lambda_setup(event, context)
-  m = util.run(bucket_name, key, params, combine)
-  util.show_duration(context, m, params)
+  util.handle(event, context, combine)
